django_apps/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.2
Django==3.1.12
django-cors-headers==3.5.0
djangorestframework==3.12.1
drf-api-tracking==1.8.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
josepy==1.5.0
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
mozilla-django-oidc==1.2.4
multitasking==0.0.9
numpy==1.20.2
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pyOpenSSL==20.0.0
python-dateutil==2.8.1
pytz==2020.1
QuantLib==1.21
requests==2.25.1
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.26.5
wrapt==1.12.1
mailchimp-marketing==3.0.36
more-itertools==8.7.0
pinax-referrals==4.0.1
pandas-market-calendars==1.6.1
